Firestore search works now... sort of. read desc.
Firestore doesn't have built in string querying which I find to be a major downside to the platform - Devs would need to add Algolia or other extension which is not free. My search implementation here is decent, and it works. It is case-sensitive. "Veg" will return "Vegetable Stew" but "veg" will not return it.